Implemented Play All using semaphores, and Add Workout layout changes

+++ app/src/main/java/xyz/cortland/fittimer/android/model/WorkoutModel.kt
@@ -10,7 +10,7 @@ data class WorkoutModel(var seconds: Int?, var workoutName: String?, var workout
     var isPlaying: Boolean? = false
     var isPlayingAll: Boolean? = false
     var isCount: Boolean? = false
-    var remainingSeconds: Long? = 0
+    var remainingSeconds: Int? = 0
     var expanded: Boolean? = false
     var isDefaultState: Boolean? = true
